Crawling/requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.3
numpy==1.23.2
openpyxl==3.0.10
pandas==1.4.4
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
soupsieve==2.3.2.post1
tornado==6.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
wincertstore==0.2
